Fix admin login build rendering
